+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- Migrated to Dart 3.0.
+- Updated example of using custom shapes.
+- Fixed animation start delay.
+
 # 1.0.0+1
 
 - Updated README.md.